🔨 Reorganize and delete comments
