+++ lib/core/my_app.dart
@@ -1,15 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'features/auth/presentation/login_page.dart';
-import 'features/home/presentation/home_page.dart';
-import 'services/auth_service.dart';
+import '../presentation/pages/login/login_page.dart';
+import '../presentation/pages/home/home_page.dart';
+import '../services/auth_service.dart';
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    // Usamos el stream de FirebaseAuth para decidir a dónde ir
     return MaterialApp(
       title: 'Flutter Auth Demo',
       theme: ThemeData(useMaterial3: true, colorSchemeSeed: Colors.blue),
