+++ lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:ilussia_tarot/app_config.dart';
-import 'firebase_options.dart'; // generado por flutterfire configure
-import 'app.dart';
+import 'package:ilussia_tarot/core/config/app_config.dart';
+import 'services/firebase_options.dart';
+import 'core/my_app.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -11,7 +11,7 @@ Future<void> main() async {
     options: DefaultFirebaseOptions.currentPlatform,
   );
   await GoogleSignIn.instance.initialize(
-    serverClientId: AppConfig.googleServerClientId,//'318316028823-lfah2772d4q5q6ql6u4edp9ugb4na5ul.apps.googleusercontent.com',
+    serverClientId: AppConfig.googleServerClientId,
   );
   
   runApp(const MyApp());
